+++ src/openmp/main.cc
@@ -1,11 +1,11 @@
 #include "../common/definition.h"
 
-double trapezoid(double x0, double xn, long long n) {
+double trapezoid(double x0, double xn, long n) {
   double h = (xn - x0) / n;
   double acum = 0;
 
 #pragma omp parallel for simd simdlen(8) reduction(+: acum)
-  for(long long i = 1; i < n; ++i) {
+  for(long i = 1; i < n; ++i) {
     acum += f(x0 + h * i);
   }
 
@@ -16,7 +16,7 @@ int main() {
   double a, b;
   a = LOWER_LIMIT;
   b = UPPER_LIMIT;
-  long long n = N;
+  long n = N;
   const double t0 = omp_get_wtime();
   double result = trapezoid(a, b, n);
   const double t1 = omp_get_wtime();
